requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-admin
flask-login
flask-wtf
email-validator
python-dotenv
faker